# Conflicts:
#	src/main/java/com/fiap/restaurant/booking/core/usecases/feedback/impl/CreateFeedBackUseCaseImpl.java
#	src/main/java/com/fiap/restaurant/booking/infrastructure/controllers/RestauranteController.java
#	src/main/java/com/fiap/restaurant/booking/infrastructure/persistence/entities/RestauranteEntity.java
#	src/test/java/com/fiap/restaurant/booking/core/usecases/restaurante/CreateRestauranteUseCaseTest.java
#	src/test/java/com/fiap/restaurant/booking/core/usecases/restaurante/FindRestauranteByCnpjUseCaseTest.java
#	src/test/java/com/fiap/restaurant/booking/core/usecases/restaurante/FindRestauranteByEnderecoBairroUseCaseTest.java
#	src/test/java/com/fiap/restaurant/booking/core/usecases/restaurante/FindRestauranteByEnderecoCidadeUseCaseTest.java
#	src/test/java/com/fiap/restaurant/booking/core/usecases/restaurante/FindRestauranteByEnderecoRuaUseCaseTest.java
#	src/test/java/com/fiap/restaurant/booking/core/usecases/restaurante/FindRestauranteByIdUseCaseTest.java
#	src/test/java/com/fiap/restaurant/booking/core/usecases/restaurante/FindRestauranteByMediaFeedbackUseCaseTest.java
#	src/test/java/com/fiap/restaurant/booking/core/usecases/restaurante/FindRestauranteByNomeUseCaseTest.java
#	src/test/java/com/fiap/restaurant/booking/core/usecases/restaurante/FindRestauranteByTipoCulinariaUseCaseTest.java
#	src/test/java/com/fiap/restaurant/booking/core/usecases/restaurante/GetAllRestaurantesUseCaseTest.java
#	src/test/java/com/fiap/restaurant/booking/utils/DefaultParamsConstants.java
#	src/test/java/com/fiap/restaurant/booking/utils/SchemaDefinitionConstants.java
#	src/test/java/com/fiap/restaurant/booking/utils/fixture/RestauranteDomainFixture.java
#	src/test/java/com/fiap/restaurant/booking/utils/fixture/RestauranteEntityFixture.java
#	src/test/resources/data.sql

+++ src/main/java/com/fiap/restaurant/booking/infrastructure/controllers/request/ReservaRequest.java
@@ -1,8 +1,8 @@
 package com.fiap.restaurant.booking.infrastructure.controllers.request;
 
 public record ReservaRequest(
-        Integer restauranteId,
-        Integer mesaId,
+        Long restauranteId,
+        Long mesaId,
         String cpfCliente,
         String dataHoraReserva
 ) {
